Fixes #26652 - Refactor folder structure of ToastList
